adding population & balance module to adjust pop and detect lag spikes, also a redundant force linking if no response from the server can be expected (no one should use this)

+++ DiscordBot/Models/Rcon/ResponseServerInfo.cs
@@ -26,7 +26,7 @@ public class ResponseServerInfo
     public string GameTime;
     
     [JsonProperty("Uptime")]
-    public int Uptime;
+    public double Uptime;
     
     [JsonProperty("Map")]
     public string Map;
